dsfjkfdsaj

+++ include/mecs-entity.h
@@ -1,6 +1,6 @@
 #ifndef MECS_ENTITY_H
 #define MECS_ENTITY_H
 
-typedef unsigned long long MecsEntity;
+typedef size_t MecsEntity;
 
 #endif
